Fixes +/-0 rating is green/red now, + shows elo on hover (#813)

+++ apis/src/common/rating_change_info.rs
@@ -2,18 +2,18 @@ use crate::responses::GameResponse;
 
 #[derive(Clone, PartialEq)]
 pub struct RatingChangeInfo {
-    pub white_rating_change: i64,
+    pub white_rating_change: f64,
     pub white_rating: u64,
-    pub black_rating_change: i64,
+    pub black_rating_change: f64,
     pub black_rating: u64,
 }
 
 impl RatingChangeInfo {
     pub fn from_game_response(gr: &GameResponse) -> Self {
         RatingChangeInfo {
-            white_rating_change: gr.white_rating_change.unwrap_or_default() as i64,
+            white_rating_change: gr.white_rating_change.unwrap_or_default(),
             white_rating: gr.white_rating.unwrap_or_default() as u64,
-            black_rating_change: gr.black_rating_change.unwrap_or_default() as i64,
+            black_rating_change: gr.black_rating_change.unwrap_or_default(),
             black_rating: gr.black_rating.unwrap_or_default() as u64,
         }
     }
